Fix FileService.DownloadFile, update example
When DownloadFile originally called client.Do(req, nil), the resp.Body was
closed and consumers of the http.Response given by
FileService.DownloadFile were unable to actually read the contents of the file.

This patch uses the client.DoAndGetReader method to return a response
with an unclosed reader. It also updates the example to count the bytes
in the file contents.